Move Qwen pixel helper to accuracy helpers
